+++ tamagui.config.ts
@@ -1,6 +1,0 @@
-import { createTamagui } from 'tamagui'
-import { config as defaultConfig } from '@tamagui/config/v3'
-
-export const tamaguiConfig = createTamagui(defaultConfig)
-
-export default tamaguiConfig